Se cambian los modelos a sus respectivos archivos de models. Se genera una configuración general del proyecto. Se corrigen ciertos errores
