+++ app.js
@@ -1,42 +1,22 @@
-import { Sequelize, DataTypes } from 'sequelize';
-import express, { json } from 'express';
+/* APP: Configuraciones principales */
+//Express configuration
+const express = require("express");
+const app = express();
 
-/* 
-CONEXIÓN A LA BASE DE DATOS
+//Body Parser
+const bodyParser = require('body-parser');
+app.use(bodyParser.urlencoded({ extended: false }));
+app.use(bodyParser.json());
+
+const PORT = 3000;
+
+
+//sequelize para las consultas
+const sequelize = require('./config/db') //Eliminar esta linea de este sitio
+
+/* Prueba para la conexión a la BD
+Eliminar después de la prueba
 */
-
-// Intento uno: (Eliminar)
-//const sequelize = new Sequelize('postgres://yhrrbfwgzphhll:dbf6ddf01cbcd368a3b7af3fd8578f36a863a71829c6f66222698eec2827f87c@ec2-35-170-146-54.compute-1.amazonaws.com:5432/da604cdd0trbqn')
-
-// Intento dos: (Eliminar)
-/* const sequelize = new Sequelize(
-    'da604cdd0trbqn',
-    'yhrrbfwgzphhll',
-    'dbf6ddf01cbcd368a3b7af3fd8578f36a863a71829c6f66222698eec2827f87c',
-    {
-        host: 'ec2-35-170-146-54.compute-1.amazonaws.com',
-        dialect: 'postgres',
-        native: true,
-        ssl: true
-    }
-); */
-
-//Intento tres: (funciona)
-const sequelize = new Sequelize(
-    {
-        database: 'da604cdd0trbqn',
-        username: 'yhrrbfwgzphhll',
-        password: 'dbf6ddf01cbcd368a3b7af3fd8578f36a863a71829c6f66222698eec2827f87c',
-        host: 'ec2-35-170-146-54.compute-1.amazonaws.com',
-        port: 5432,
-        ssl: true,
-        dialect: 'postgres',
-        dialectOptions: {
-            "ssl": { "rejectUnauthorized": false }
-        }
-    }
-);
-
 //Conexión a la base de datos con try/catch
 try {
     sequelize.authenticate();
@@ -45,210 +25,12 @@ try {
     console.error('Hubo un problema con la conexión', error);
 }
 
-/* 
-CREACIÓN DE MODELOS
-*/
+const Alumno = require('./models/Alumno') //Eliminar esta linea de este sitio
+const Maestro = require('./models/Maestro') //Eliminar esta linea de este sitio
+const Actividad = require('./models/Actividad') //Eliminar esta linea de este sitio
+const Inscripcion = require('./models/Inscripcion') //Eliminar esta linea de este sitio
+const Resena = require('./models/Resena') //Eliminar esta linea de este sitio
 
-//Alumno
-const Alumno = sequelize.define('Alumno', {
-    id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        primaryKey: true,
-        autoIncrement: true
-    },
-    nombre: {
-        type: DataTypes.TEXT,
-        allowNull: false
-    },
-    apellido: {
-        type: DataTypes.TEXT,
-        allowNull: false
-    },
-    email: {
-        type: DataTypes.TEXT,
-        allowNull: false
-    },
-    password: {
-        type: DataTypes.TEXT,
-        allowNull: false
-    }
-}, {
-    freezeTableName: true,
-    timestamps: false
-});
-
-// Maestro
-const Maestro = sequelize.define('Maestro', {
-    id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        primaryKey: true,
-        autoIncrement: true
-    },
-    nombre: {
-        type: DataTypes.TEXT,
-        allowNull: false
-    },
-    apellido: {
-        type: DataTypes.TEXT,
-        allowNull: false
-    },
-    email: {
-        type: DataTypes.TEXT,
-        allowNull: false
-    },
-    password: {
-        type: DataTypes.TEXT,
-        allowNull: false
-    }
-}, {
-    freezeTableName: true,
-    timestamps: false
-});
-
-//Actividad
-const Actividad = sequelize.define('Actividad', {
-    id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        primaryKey: true,
-        autoIncrement: true
-    },
-    maestro_id: {
-        type: DataTypes.INTEGER,
-        allowNull: false
-    },
-    nombre: {
-        type: DataTypes.TEXT,
-        allowNull: false
-    },
-    costo: {
-        type: DataTypes.INTEGER,
-        allowNull: false
-    },
-    descripcion: {
-        type: DataTypes.TEXT,
-        allowNull: false
-    },
-    dias: {
-        type: DataTypes.TEXT,
-        allowNull: false
-    },
-    horario: {
-        type: DataTypes.TEXT,
-        allowNull: false
-    }
-}, {
-    freezeTableName: true,
-    timestamps: false
-});
-
-//Inscripción
-const Inscripcion = sequelize.define('Inscripcion', {
-    id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        primaryKey: true,
-        autoIncrement: true
-    },
-    alumno_id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-    },
-    actividad_id: {
-        type: DataTypes.INTEGER,
-        allowNull: false
-    }
-}, {
-    freezeTableName: true,
-    timestamps: false
-});
-
-//Reseña
-const Resena = sequelize.define('Resena', {
-    id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        primaryKey: true,
-        autoIncrement: true
-    },
-    alumno_id: {
-        type: DataTypes.INTEGER,
-        allowNull: false
-    },
-    actividad_id: {
-        type: DataTypes.INTEGER,
-        allowNull: false
-    },
-    calificacion: {
-        type: DataTypes.INTEGER,
-        allowNull: false
-    },
-    comentario: {
-        type: DataTypes.TEXT,
-        allowNull: false
-    }
-}, {
-    freezeTableName: true,
-    timestamps: false
-});
-
-/* 
-ASOCIACIONES
-*/
-
-//Alumno
-Alumno.hasMany(Inscripcion, {
-    foreignKey: 'alumno_id'
-})
-
-Alumno.hasMany(Resena, {
-    foreignKey: 'alumno_id'
-})
-
-//Maestro
-Maestro.hasMany(Actividad, {
-    foreignKey: 'maestro_id'
-})
-
-//Actividad
-Actividad.hasMany(Inscripcion, {
-    foreignKey: 'actividad_id'
-})
-
-Actividad.hasMany(Resena, {
-    foreignKey: 'actividad_id'
-})
-
-Actividad.belongsTo(Maestro, {
-    foreignKey: 'maestro_id'
-})
-
-//Inscripción
-Inscripcion.belongsTo(Alumno, {
-    foreignKey: 'alumno_id'
-})
-
-Inscripcion.belongsTo(Actividad, {
-    foreignKey: 'actividad_id'
-})
-
-//Reseña
-Resena.belongsTo(Alumno, {
-    foreignKey: 'alumno_id'
-})
-
-Resena.belongsTo(Actividad, {
-    foreignKey: 'actividad_id'
-})
-
-await sequelize.sync();
-
-//Express configuration
-const app = express();
-const PORT = 3000;
-app.use(json());
 
 //Consulta general
 app.get('/alumnos', async(req, res) => {
@@ -304,3 +86,6 @@ app.get('/alumnos/:nombre', async(req, res) => { //id
 app.listen(PORT, () => {
     console.log(`App listening on port ${PORT}`);
 });
+
+
+
