+++ models/Maestro.js
@@ -1,0 +1,36 @@
+const { Sequelize, DataTypes, Op } = require('sequelize');
+const sequelize = require('../config/db')
+
+const Actividad = require('./Actividad')
+
+// Maestro
+const Maestro = sequelize.define('Maestro', {
+    id: {
+        type: DataTypes.INTEGER,
+        allowNull: false,
+        primaryKey: true,
+        autoIncrement: true
+    },
+    nombre: {
+        type: DataTypes.TEXT,
+        allowNull: false
+    },
+    apellido: {
+        type: DataTypes.TEXT,
+        allowNull: false
+    },
+    email: {
+        type: DataTypes.TEXT,
+        unique: true,
+        allowNull: false
+    },
+    password: {
+        type: DataTypes.TEXT,
+        allowNull: false
+    }
+}, {
+    freezeTableName: true,
+    timestamps: false
+});
+
+module.exports = Maestro;
